+++ CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Changed
+- Add the missing `type` column into the FeatureFlagEntity.
+- Modify the code to verify if the metadata feature is enabled.
+
 ## [0.1.0] 2024-09-29
 
+### Added
 - Add Slack Notifier
 - Support Whitelisting and Blacklisting for UserTargeting in Metadata